fix PR Commengts
